+++ src/components/send/confirm/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import connect from 'redux-connect-decorator';
 import { View, Image, Platform } from 'react-native';
 import { validatePassphrase } from '../../../utilities/passphrase';
-import { extractPublicKey } from '../../../utilities/api/account';
+import { extractPublicKey } from '../../../utilities/api/lisk/account';
 import Input from '../../toolBox/input';
 import { P } from '../../toolBox/typography';
 import { IconButton } from '../../toolBox/button';
