+++ src/utilities/passphrase.js
@@ -2,7 +2,7 @@ import Lisk from '@liskhq/lisk-client';
 import * as Keychain from 'react-native-keychain';
 import FingerprintScanner from 'react-native-fingerprint-scanner';
 import { Platform } from 'react-native';
-import { extractAddress } from './api/account';
+import { extractAddress } from './api/lisk/account';
 
 const fullWordsList = Lisk.passphrase.Mnemonic.wordlists.EN;
